requirements.txt
appnope==0.1.2
argon2-cffi==21.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
configparser==5.1.0
crayons==0.4.0
cryptography==35.0.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
fonttools==4.28.2
h11==0.12.0
idna==3.3
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.3
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lxml==4.6.4
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
mysql-connector-python==8.0.27
nbclient==0.5.8
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.4
outcome==1.1.0
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.10.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.2.0
QtPy==1.11.2
requests==2.26.0
scipy==1.7.3
seaborn==0.11.2
selenium==4.0.0
Send2Trash==1.8.0
setuptools-scm==6.3.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
terminado==0.12.1
testpath==0.5.0
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wcwidth==0.2.5
webdriver-manager==3.5.2
webencodings==0.5.1
widgetsnbextension==3.5.2
wsproto==1.0.0
